CropAnalytics/requirements.txt
# Core Deep Learning
tensorflow==2.15.0
numpy==1.24.3
pandas==2.1.0
Pillow==10.0.0

# Web Framework
streamlit==1.28.0
streamlit-folium==0.15.0

# Geolocation & Mapping
folium==0.14.0

# Image Processing
opencv-python==4.8.1.78
scikit-image==0.21.0

# Utilities
matplotlib==3.7.2
scikit-learn==1.3.0
